Test window resize hook
