+++ package.json
@@ -6,6 +6,7 @@
     "@material-ui/core": "^4.10.2",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.56",
+    "@rehooks/component-size": "^1.0.3",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
@@ -37,6 +38,7 @@
     "redux-devtools-extension": "^2.13.8",
     "remotedev-server": "^0.3.1",
     "reselect": "^4.0.0",
+    "resize-observer-polyfill": "^1.5.1",
     "styled-components": "^5.1.1",
     "typescript": "3.8.3"
   },
